+++ examples/transports.js
@@ -2,15 +2,24 @@
 
 let Papertrail = require('@mnm/logger-papertrail')
 let SlackLogger = require('@mnm/logger-slack')
-let log = require('..')('logger')
 let environment = require('envobj')
 let Pretty = require('bistre')
+let log = require('..')('jack:bot')
 let Log = require('..')
+
+
+/**
+ * Required environment variables
+ */
 
 let env = environment({
   PAPERTRAIL_URL: '',
   SLACK_WEBHOOK: ''
 })
+
+/**
+ * Logging
+ */
 
 if (env.PAPERTRAIL_URL) {
   let papertrail = Papertrail(env.PAPERTRAIL_URL)
@@ -23,14 +32,14 @@ if (env.PAPERTRAIL_URL) {
 
 // log to slack
 if (env.SLACK_WEBHOOK) {
-  let slack_logger = SlackLogger(env.SLACK_WEBHOOK, {
-    username: 'Sutra'
+  let jackops_logger = SlackLogger(env.SLACK_WEBHOOK, {
+    icon_url: 'https://cldup.com/8KHsv7mkQW.png',
+    username: 'Jackops'
   })
-  Log('to:slack').pipe(slack_logger)
-  log.info.pipe(slack_logger)
-  log.warn.pipe(slack_logger)
-  log.error.pipe(slack_logger)
-  log.fatal.pipe(slack_logger)
+  Log('to:slack').pipe(jackops_logger)
+  log.warn.pipe(jackops_logger)
+  log.error.pipe(jackops_logger)
+  log.fatal.pipe(jackops_logger)
 }
 
 // default logging
@@ -40,11 +49,16 @@ log.warn.pipe(pretty)
 log.error.pipe(pretty)
 log.fatal.pipe(pretty)
 pretty.pipe(process.stderr)
-
 let i = 0
 let team = { a: 1, b: 2, c: 3}
 log.info('hi there', { team: team })
 let sid = setInterval(function() {
-  log.info('hi there', { team: team })
+  let team_id = 'team'
+  let message = {
+    user: 'matt',
+    text: 'lol!'
+  }
+  Log('to:slack').info('*%s (%s)* ← %s', message.user, team_id, message.text)
+  log.info('%s (%s) ← %s', message.user, team_id, message.text)
 }, 1000)